new gun, cubescript aliases, bug fix, and new HUD
There's a new HUD to play with, but it is a bit iffy. Go into edit mode
to activate it.
A new gun has been added. It's a machine gun. It uses the plasma gun
model for now.
More cubescript aliases for frags, deaths and checking for the player
crouching, jumping or being in water.
A bug fix for ammo being infinite has been patched.