new map, and addons!
I created a new CtF map called skytowers, and I have created an addons
system. I will write a guide on how to create addons.
I have include three test addons: "Nimbassa", which overwrites the music
player to play Nimbassa Core by plasterbrain and that only (and, of
course, a way of reversing it), "Notepad", which allows to access the
notepad program inside of Tesseract, and "Press 8", which is just a joke
stolen from The Stanley Parable. These mods will not be in the finished
release.